requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
numpy==2.1.3
Pillow==10.4.0
requests==2.32.3
python-multipart==0.0.12
